# translation-layer/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "translation-layer",
        "description": "",
        "contact": {
            "name": "Space and Time Labs"
        },
        "license": {
            "name": "MIT-0",
            "identifier": "MIT-0"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/mainnet/create_table": {
            "post": {
                "tags": [
                    "create-table"
                ],
                "summary": "Submits a transaction to create a new table in the indexing system.",
                "description": "This endpoint constructs a table creation request and submits it as an extrinsic to the blockchain.\n\n# Request Body\n- `schema_name` (String): The schema namespace for the table.\n- `ddl_statement` (String): The Data Definition Language (DDL) statement defining the table structure.\n- `source` (String): The source blockchain network.\n- `mode` (String): The indexing mode for the table.\n- `tables` (Vec<TableRequest>): A list of tables to be created.\n- `table_name` (String): The name of the table.\n\n# Responses\n- **200 OK**: Table successfully created.\n- **400 BAD REQUEST**: Invalid request parameters.\n- **500 INTERNAL SERVER ERROR**: Transaction submission failed.\n\n# Example Usage\n```sh\ncurl -X POST \"http://127.0.0.1:3000/create_table\" -H \"Content-Type: application/json\" -d '{\n    \"schemaName\": \"public\",\n    \"ddlStatement\": \"CREATE TABLE example_table (id SERIAL PRIMARY KEY, data TEXT);\",\n    \"source\": \"Ethereum\",\n    \"mode\": \"indexing\",\n    \"tables\": [\n        {\n            \"tableName\": \"example_table\",\n            \"schemaName\": \"public\",\n            \"ddlStatement\": \"CREATE TABLE example_table (id SERIAL PRIMARY KEY, data TEXT);\"\n        }\n    ],\n    \"tableName\": \"example_table\"\n}'\n```",
                "operationId": "create_table",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTableRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Table created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mainnet/drop_table": {
            "post": {
                "tags": [
                    "drop-table"
                ],
                "summary": "Submits a transaction to drop a table from the indexing system.",
                "description": "This endpoint allows for the removal of an indexed table from the blockchain storage.\n\n# Request Body\n- `schema_name` (String): The schema namespace of the table.\n- `table_name` (String): The name of the table to be removed.\n- `source` (String): The source blockchain network.\n- `mode` (String): The indexing mode.\n\n# Responses\n- **200 OK**: Table successfully removed.\n- **400 BAD REQUEST**: Invalid request parameters.\n- **500 INTERNAL SERVER ERROR**: Transaction submission failed.\n\n# Example Usage\n```sh\ncurl -X POST \"http://127.0.0.1:3000/drop_table\" -H \"Content-Type: application/json\" -d '{\n    \"schemaName\": \"public\",\n    \"tableName\": \"example_table\",\n    \"source\": \"Ethereum\",\n    \"mode\": \"indexing\"\n}'\n```",
                "operationId": "drop_table",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DropTableRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Table dropped successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/testnet/create_table": {
            "post": {
                "tags": [
                    "create-table"
                ],
                "summary": "Submits a transaction to create a new table in the indexing system.",
                "description": "This endpoint constructs a table creation request and submits it as an extrinsic to the blockchain.\n\n# Request Body\n- `schema_name` (String): The schema namespace for the table.\n- `ddl_statement` (String): The Data Definition Language (DDL) statement defining the table structure.\n- `source` (String): The source blockchain network.\n- `mode` (String): The indexing mode for the table.\n- `tables` (Vec<TableRequest>): A list of tables to be created.\n- `table_name` (String): The name of the table.\n\n# Responses\n- **200 OK**: Table successfully created.\n- **400 BAD REQUEST**: Invalid request parameters.\n- **500 INTERNAL SERVER ERROR**: Transaction submission failed.\n\n# Example Usage\n```sh\ncurl -X POST \"http://127.0.0.1:3000/create_table\" -H \"Content-Type: application/json\" -d '{\n    \"schemaName\": \"public\",\n    \"ddlStatement\": \"CREATE TABLE example_table (id SERIAL PRIMARY KEY, data TEXT);\",\n    \"source\": \"Ethereum\",\n    \"mode\": \"indexing\",\n    \"tables\": [\n        {\n            \"tableName\": \"example_table\",\n            \"schemaName\": \"public\",\n            \"ddlStatement\": \"CREATE TABLE example_table (id SERIAL PRIMARY KEY, data TEXT);\"\n        }\n    ],\n    \"tableName\": \"example_table\"\n}'\n```",
                "operationId": "create_table",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTableRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Table created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/testnet/drop_table": {
            "post": {
                "tags": [
                    "drop-table"
                ],
                "summary": "Submits a transaction to drop a table from the indexing system.",
                "description": "This endpoint allows for the removal of an indexed table from the blockchain storage.\n\n# Request Body\n- `schema_name` (String): The schema namespace of the table.\n- `table_name` (String): The name of the table to be removed.\n- `source` (String): The source blockchain network.\n- `mode` (String): The indexing mode.\n\n# Responses\n- **200 OK**: Table successfully removed.\n- **400 BAD REQUEST**: Invalid request parameters.\n- **500 INTERNAL SERVER ERROR**: Transaction submission failed.\n\n# Example Usage\n```sh\ncurl -X POST \"http://127.0.0.1:3000/drop_table\" -H \"Content-Type: application/json\" -d '{\n    \"schemaName\": \"public\",\n    \"tableName\": \"example_table\",\n    \"source\": \"Ethereum\",\n    \"mode\": \"indexing\"\n}'\n```",
                "operationId": "drop_table",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DropTableRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Table dropped successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponse": {
                "type": "object",
                "description": "Generic API response structure.",
                "required": [
                    "success"
                ],
                "properties": {
                    "errMsg": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "An optional error message in case of failure."
                    },
                    "success": {
                        "type": "boolean",
                        "description": "Indicates whether the request was successful."
                    },
                    "txHash": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The transaction hash associated with the request, if applicable."
                    }
                }
            },
            "CommitmentScheme": {
                "type": "string",
                "enum": [
                    "HyperKzg",
                    "DynamicDory"
                ]
            },
            "CreateTableRequest": {
                "type": "object",
                "description": "Represents a request to create tables for storing indexed data.",
                "required": [
                    "schemaName",
                    "ddlStatement",
                    "tables",
                    "tableName",
                    "tableType"
                ],
                "properties": {
                    "ddlStatement": {
                        "type": "string",
                        "description": "Schema level DDL statement"
                    },
                    "schemaName": {
                        "type": "string",
                        "description": "Schema name"
                    },
                    "tableName": {
                        "type": "string",
                        "description": "The table name."
                    },
                    "tableType": {
                        "$ref": "#/components/schemas/TableType",
                        "description": "The table type"
                    },
                    "tables": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TableRequest"
                        },
                        "description": "A list of tables to be created."
                    }
                }
            },
            "DropTableRequest": {
                "type": "object",
                "description": "Represents a request to drop a table.",
                "required": [
                    "schemaName",
                    "tableName",
                    "source",
                    "mode",
                    "tableType"
                ],
                "properties": {
                    "mode": {
                        "type": "string",
                        "description": "The indexing mode."
                    },
                    "schemaName": {
                        "type": "string",
                        "description": "The schema name in which the table exists."
                    },
                    "source": {
                        "type": "string",
                        "description": "The source blockchain network."
                    },
                    "tableName": {
                        "type": "string",
                        "description": "The table name to drop."
                    },
                    "tableType": {
                        "$ref": "#/components/schemas/TableType",
                        "description": "Table type"
                    }
                }
            },
...
            "TableRequest": {
                "type": "object",
                "description": "Represents an individual table entry inside a `CreateTableRequest`.",
                "required": [
                    "tableType",
                    "tableName",
                    "schemaName",
                    "ddlStatement",
                    "commitment",
                    "snapshotUrl",
                    "commitmentScheme"
                ],
                "properties": {
                    "commitment": {
                        "type": "string",
                        "description": "commitment data"
                    },
                    "commitmentScheme": {
                        "$ref": "#/components/schemas/CommitmentScheme",
                        "description": "commitment scheme"
                    },
                    "ddlStatement": {
                        "type": "string",
                        "description": "The Data Definition Language (DDL) statement defining tnkhe table schema."
                    },
                    "schemaName": {
                        "type": "string",
                        "description": "The namespace of the table."
                    },
                    "snapshotUrl": {
                        "type": "string",
                        "description": "snapshot location"
                    },
                    "tableName": {
                        "type": "string",
                        "description": "The table name."
                    },
                    "tableType": {
                        "$ref": "#/components/schemas/TableType",
                        "description": "table type"
                    }
                }
            },
            "TableType": {
                "type": "string",
                "description": "Represents the type of a table being created or referenced within the translation layer.\n\nThis enum categorizes tables based on their intended usage and ownership model.\nIt is used to determine how data should be interpreted, stored, or indexed\nacross different subsystems.\n\n# Variants\n- `CoreBlockchain`: A default system table directly tied to the core blockchain state.\n- `SCI`: A table used for indexing smart contract events and data (Smart Contract Indexing).\n- `Community`: A table managed or curated by the community or external contributors.",
                "enum": [
                    "CoreBlockchain",
                    "SCI",
                    "Community"
                ]
            },
...
        }
    }
}